requirements.txt
python-telegram-bot
python-telegram-bot[job-queue]
redis
pillow
Python-IO
asyncio
pymongo
python-dotenv
fastapi 
uvicorn
pybase64